refactor(Itinerary): Decompose ItineraryViewer into modular helpers and widgets
This commit refactors the `ItineraryViewer` to improve separation of concerns by extracting logic into dedicated helper classes and standalone widgets.

- **Timeline Helpers**: Introduced `TimelineEventFactory`, `TimelineEventFormatter`, and `TimelineThemeHelper` to encapsulate event creation, data formatting, and styling logic.
- **`TimelineRebuildHelper`**: Extracted state comparison logic to efficiently manage widget rebuilds based on entity updates.
- **`TimelineItemWidget`**: Moved the timeline card UI into a separate widget with dedicated styling constants.
- **`ItineraryViewer`**: Simplified the main widget to coordinate these new helpers, removing monolithic logic.
- **Dependencies**: Updated `meta` and `test_api` packages.
